Separated queries from Model.php to Query.php
